% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_data.R
\name{create_partition.default}
\alias{create_partition.default}
\title{Create Partition}
\usage{
\method{create_partition}{default}(df, time, lon, lat)
}
\arguments{
\item{df}{a data frame}

\item{time}{time of event}

\item{lon}{lon of event}

\item{lat}{lat of event}
}
\value{
partitions
}
\description{
Create Partition
}
