% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_data.R
\name{create_partition.sf}
\alias{create_partition.sf}
\title{Create a list of partitioned data frames for lm models}
\usage{
\method{create_partition}{sf}(df, n_par = 10, centre, geometry_col)
}
\arguments{
\item{df}{data frame with a}

\item{n_par}{the number of partitions to create, defaults to 10}
}
\value{
list of data frames (length of n_par)
}
\description{
Create a list of partitioned data frames for lm models
}
\examples{

}
