% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_data.R
\name{create_plane}
\alias{create_plane}
\title{Create/simulate a plane with 3 dimensinos and a value level}
\usage{
create_plane(n_points = 500)
}
\arguments{
\item{n_points}{The number of points to simulate}
}
\value{
tbl_df
}
\description{
Create/simulate a plane with 3 dimensinos and a value level
}
