% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_data.R
\name{cut_circle}
\alias{cut_circle}
\title{Cut the circle into a list of useable tibbles}
\usage{
cut_circle(df, cuts = 5, point = c(0, 0), pb = NULL)
}
\arguments{
\item{point}{}
}
\description{
Cut the circle into a list of useable tibbles
}
\details{
The variable that should be used in modelling is distance along
with anyany additional predictors in the original data frame.

These should be given in formula notation to the modelling functions.
}
