% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dist3d}
\alias{dist3d}
\title{Distance from point to line}
\usage{
dist3d(a, b, c)
}
\arguments{
\item{a}{point}

\item{b}{line point 1}

\item{c}{line point 2}
}
\description{
Distance from point to line
}
\details{
The points b and c have a line travelling through them.
This function calculates the distance from  point a to this line.

All arguments must be  of the form c(x, y, z)
}
