% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_data.R
\name{mark_lines}
\alias{mark_lines}
\title{Mark the lines that will be used to cut the circle}
\usage{
mark_lines(n_lines = 5, centre = c(0, 0), interval = 0.5)
}
\arguments{
\item{centre}{}
}
\value{

}
\description{
Mark the lines that will be used to cut the circle
}
\details{
This function generates a set of lines that divide a circle and will
be used to make predictions. points will eventually be weighted based on their
}
