% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_data.R
\name{mark_planes}
\alias{mark_planes}
\title{Mark planes evenly around a spherical (or elipsoid) object}
\usage{
mark_planes(n_planes, centre = c(0, 0, 0))
}
\arguments{
\item{centre}{}
}
\description{
Mark planes evenly around a spherical (or elipsoid) object
}
\examples{
mark_planes(5)
}
