% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-partitions.R
\name{model_partitions}
\alias{model_partitions}
\title{Model the partitions that have been made}
\usage{
model_partitions(lst, optim_span = FALSE)
}
\arguments{
\item{lst}{A list of lists. The first level represents the planar cuts,
 the second represents the linear cuts along the line.

 @details
 This is used to create the models from the output of the}
}
\description{
Model the partitions that have been made
}
