% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{util_abline}
\alias{util_abline}
\title{Utility abline}
\usage{
util_abline(ls, value = "value", ...)
}
\arguments{
\item{...}{}
}
\description{
This is a utility function that makes it easier to plot the results of mark_lines() function
}
